% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_params}
\alias{add_params}
\title{Add non-NULL parameters to a parameter list}
\usage{
add_params(params, new_params)
}
\arguments{
\item{params}{Existing parameter list}

\item{new_params}{Named list of new parameters to add if non-NULL}
}
\value{
Updated parameter list
}
\description{
Add non-NULL parameters to a parameter list
}
\keyword{internal}
