% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{clear_column_prefix}
\alias{clear_column_prefix}
\title{Clean column names by removing table prefixes}
\usage{
clear_column_prefix(df)
}
\arguments{
\item{df}{Data frame with column names to clean}
}
\value{
Data frame with cleaned column names
}
\description{
Clean column names by removing table prefixes
}
