% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_timestamp}
\alias{format_timestamp}
\title{Format a timestamp for the SMEAR API (ISO 8601 in UTC+2)}
\usage{
format_timestamp(time)
}
\arguments{
\item{time}{POSIXt object to format}
}
\value{
Formatted timestamp string
}
\description{
Format a timestamp for the SMEAR API (ISO 8601 in UTC+2)
}
\keyword{internal}
