% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_api_url}
\alias{get_api_url}
\title{Get the base API URL from environment or default}
\usage{
get_api_url()
}
\value{
The base API URL
}
\description{
Get the base API URL from environment or default
}
\keyword{internal}
