% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{get_stations}
\alias{get_stations}
\title{Get all SMEAR stations}
\usage{
get_stations()
}
\value{
A data frame containing information about all stations
}
\description{
Convenience function to get information about all SMEAR stations
}
