% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_tables}
\alias{get_tables}
\title{Get All SMEAR Tables}
\usage{
get_tables()
}
\value{
A data frame containing information about all tables
}
\description{
Convenience function to get a list of all available SMEAR tables
}
