% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_query_list}
\alias{make_query_list}
\title{Create a list with the same keyname for each value}
\usage{
make_query_list(keyname, variables)
}
\arguments{
\item{variables}{List of variables}

\item{params}{The key to use}
}
\value{
Updated parameter list
}
\description{
Create a list with the same keyname for each value
}
\keyword{internal}
