% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{search_stations}
\alias{search_stations}
\title{Search SMEAR Stations}
\usage{
search_stations(station = NULL, name = NULL)
}
\arguments{
\item{station}{Optional station name to filter results (e.g., "HYY")}

\item{name}{Optional station's human readable name to filter results (e.g., "Hyytiälä")}
}
\value{
A data frame containing station information including id, name, location, and other metadata
}
\description{
Get information about SMEAR measurement stations
}
